+++ src/Documentation/Decorators/EndpointDecorators.ts
@@ -1,4 +1,4 @@
-import { RouteDecorator } from "../../Router/RouteTypes";
+import { RouteDecorator } from "../../Router";
 import Route from "../../Router/Route";
 import MetadataStore from "../MetadataStore";
 import { IEndpointDocumentation } from "../Definition/RouteMetadata";